style(...): 폴더링 방식 변경 및 네이밍 변경
