+++ src/main/java/roomescape/Time/controller/dto/ReservationTimeRequest.java
@@ -1,0 +1,6 @@
+package roomescape.Time.controller.dto;
+
+import java.time.LocalTime;
+
+public record ReservationTimeRequest(LocalTime startAt) {
+}
