+++ src/main/java/roomescape/Time/controller/dto/ReservationTimeResponse.java
@@ -1,0 +1,6 @@
+package roomescape.Time.controller.dto;
+
+import java.time.LocalTime;
+
+public record ReservationTimeResponse(long id, LocalTime startAt) {
+}
