+++ src/main/java/roomescape/Time/dao/TimeDao.java
@@ -1,0 +1,15 @@
+package roomescape.Time.dao;
+
+import java.util.List;
+import roomescape.Time.domain.Time;
+
+public interface TimeDao {
+
+    void save(Time reservationTime);
+
+    List<Time> findAll();
+
+    Time findById(long reservationTimeId);
+
+    void deleteById(long reservationTimeId);
+}
