--- src/main/java/roomescape/controller/AdminController.java
@@ -1,24 +1,0 @@
-package roomescape.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-@RequestMapping("/admin")
-public class AdminController {
-    @GetMapping
-    public String mainPage() {
-        return "admin/index";
-    }
-
-    @GetMapping("/reservation")
-    public String reservationPage() {
-        return "admin/reservation";
-    }
-
-    @GetMapping({"/time"})
-    public String timePage() {
-        return "admin/time";
-    }
-}
