--- src/main/java/roomescape/controller/TimeViewController.java
@@ -1,12 +1,0 @@
-package roomescape.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class TimeViewController {
-    @GetMapping("/time")
-    public String timePage() {
-        return "admin/time";
-    }
-}
