--- src/main/java/roomescape/controller/dto/ReservationRequest.java
@@ -1,6 +1,0 @@
-package roomescape.controller.dto;
-
-import java.time.LocalDate;
-
-public record ReservationRequest(LocalDate date, String name, long timeId) {
-}
