--- src/main/java/roomescape/controller/dto/ReservationTimeRequest.java
@@ -1,6 +1,0 @@
-package roomescape.controller.dto;
-
-import java.time.LocalTime;
-
-public record ReservationTimeRequest(LocalTime startAt) {
-}
