--- src/main/java/roomescape/controller/dto/ReservationTimeResponse.java
@@ -1,6 +1,0 @@
-package roomescape.controller.dto;
-
-import java.time.LocalTime;
-
-public record ReservationTimeResponse(long id, LocalTime startAt) {
-}
