--- src/main/java/roomescape/dao/ReservationDaoImpl.java
@@ -1,13 +1,0 @@
-package roomescape.dao;
-
-import java.util.List;
-import roomescape.domain.Reservation;
-
-public interface ReservationDaoImpl {
-
-    void save(Reservation reservation);
-
-    List<Reservation> findAll();
-
-    void deleteById(long reservationId);
-}
