+++ src/main/java/roomescape/reservation/controller/dto/ReservationResponse.java
@@ -1,0 +1,7 @@
+package roomescape.reservation.controller.dto;
+
+import java.time.LocalDate;
+import roomescape.Time.domain.Time;
+
+public record ReservationResponse(long id, String name, LocalDate date, Time time) {
+}
