+++ src/main/java/roomescape/reservation/dao/ReservationDao.java
@@ -1,0 +1,13 @@
+package roomescape.reservation.dao;
+
+import java.util.List;
+import roomescape.reservation.domain.Reservation;
+
+public interface ReservationDao {
+
+    void save(Reservation reservation);
+
+    List<Reservation> findAll();
+
+    void deleteById(long reservationId);
+}
