+++ src/main/java/roomescape/view/controller/AdminViewController.java
@@ -1,0 +1,24 @@
+package roomescape.view.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequestMapping("/admin")
+public class AdminViewController {
+    @GetMapping
+    public String mainPage() {
+        return "admin/index";
+    }
+
+    @GetMapping("/reservation")
+    public String reservationPage() {
+        return "admin/reservation";
+    }
+
+    @GetMapping({"/time"})
+    public String timePage() {
+        return "admin/time";
+    }
+}
